feat(core-module): add core && fixbugs logic author

+++ auth-service/src/main/java/com/r2s/auth/dto/AuthResponseDTO.java
@@ -10,7 +10,7 @@ public class AuthResponseDTO {
 
     @Builder.Default
     private final String tokenType = "Bearer";
-    private final Long expiresIn;
+    private final int expiresIn;
     private final String username;
     private final String role;
